Make the text appearance consistant in add vpn dialog
Set the password fields text appearance programatically to get the
appearance consistant.

The style set in xml will be overriden if the field is set to password.

I set the text appearance programmatically in order to keep the
appearance consistent with the others.  The color of the text become
gray after setting the text appearance for some reason.

Note that I also do changes to the password field which is not mentioned
in the bug.

Test: visual inspection.
Bug: 127728228
Change-Id: Ie206f60877eb39addc1370ae4e2aee379597960e

Former-commit-id: e0f24e6f5e922e84d81510cea8833fa23157efb8