WifiCallingSettingsForSubTest: Solve casting problem in setup
Solves casting problem in WifiCallingSettingsForSubTest that causes all
tests in this file to fail in the setup-function.

The wfc mode buttons has changed from being a ListPreference to be a
ListWithEntrySummaryPreference. The test cases have now been updated
accordingly.

Problem was introduced by the following two patches that conflict with
each other. The test cases in WifiCallingSettingsForSubTest works
correctly when each patch is tested in isolation, but failed when they
were both merged:
25d8ab05ef817eca30b47ae292be7ddf029071d5 [formerly 73ffcf49990ab3134b139607c0225f6ae1646c18]
57b8a22d4046b686b4002c77f1daca8e820dde16 [formerly 8498436a9613c806d4e1085074aea22f1a78675a]

Bug: 129545431
Test: make RunSettingsRoboTests \
ROBOTEST_FILTER=WifiCallingSettingsForSubTest
Change-Id: Ib6988e514bbad2023ba6cfc2e1d8734b68cabc31

Former-commit-id: 428ba7cde651273696021d4a7269351595304023