Use better string in dialog asking which SIM to use for SMS
Right now when you have a dual-SIM device but no default set for which
one to use for SMS, when you go to try and send an SMS you'll get a
dialog asking which SIM to use with the generic string of "Select a SIM
card". We recently added a more specific string for this in a separate
CL, which reads "Select a SIM card for SMS". This CL switches the code
to use that more specific string.

Bug: 129021763
Test: compile / manual
Change-Id: I7d12a7191de23366b0a6b6888021e57ff9ddced9

Former-commit-id: 85717d9f8d17a153fe3e7c13d24e7f7b205cdcdd