Get rid of extra whitespace at top of Network & internet page
In DSDS mode, we put an extra header at the top of the Network &
internet page with a listing of your available connectivity (mobile
subscriptions, plus wifi). The PreferenceCategory that contains these
prefs has space reserved for a title by default, but we don't need or
want a title here. This CL gets rid of that extra space by using the
custom layout we already have and use in other places for this purpose.

Fixes: 129562555
Test: visual (get device into DSDS mode with multiple SIMs, and look at
the header at the top of the Network & internet page - there should be
no extra space at the top)

Change-Id: I8f9aae9c07303b2d7ecfc8b6722a4d66555392f1

Former-commit-id: 997cabb8a352c2fca0c1f71ac4525deea0246e4f