Make RadioInfo settings menu work for mSIM
Just add a spinner to the top of the screen which lets you select the
phoneIndex. Upating this value updates the mTelephonyManager
mImsManager, and mPhoneStateListener to use the selected phoneIndex and
the subscription associated with it.

Also adds fields for current subId and default data sim subId.

Test: manual test
Fixes: 128033739
Change-Id: Id075e3bffe20523d8ceeb8f2e4320f995abdab55
Merged-In: Id075e3bffe20523d8ceeb8f2e4320f995abdab55

Former-commit-id: b69e8f2615291ad93a52a974d501b18fe8e7b388