Creates user_switcher_enabled setting to match setting
When the user enters UserSettings, the switch displays the setting as
on, but the setting is not created until the user modifies this for the
first time. This change makes sure that the setting is created when the
user enters UserSettings for the first time.

Fixes: 130270878
Test: manual, switch to second user and observe switcher icons.
Change-Id: I30bc5025c178da3668a76c828c34545c1d98b085
(cherry picked from commit 11bdddece4446683f68c4f74f0528dc746053e0c [formerly a52961d80352fd20edcbb5308c13a7167096c4ac])

Former-commit-id: aaa7aef9845f0012ed6724db6ec718aca2ace10a