Update intent in SimNotification
When you add/remove a SIM, telephony notifies settings and we pop up a
notification. Tapping on that notification was still leading to our old
SIM cards page, but should instead lead to the Network & internet page
where we've added a bunch of new UI for doing the same sorts of
operations.

Bug: 128859223
Test: manual (remove a SIM or add a SIM in DSDS mode; a notification
telling you that sim cards have changed should pop up, clicking that
should lead to Network & internet page in settings)

Change-Id: I9c1e656dcc3e8fc9147bb99861d8fdc15ed09be9

Former-commit-id: 8eb1660da74c11195719a43838bd3daa0e90d899