Implement off state for SIM details page
When a subscription is disabled, we need to hide most things on the
details page. The UX mocks called for leaving in a few things such as
the data usage header and some carrier information, but in tests I've
done it looks like the telephony APIs don't have any of that available
when the subscription is disabled. So for now the only thing left on the
page in the disabled case is the on/off toggle at the very top.

Fixes: 129665956
Test: make RunSettingsRoboTests
Change-Id: I98f4590fca38d03b1e5ff2cf0d2b4876d2adf936

Former-commit-id: 74d8d9c31f9b0eca571f7fa7c3293207dd5afcdb