Update Battery Saver schedule to add order to seekbar
If the first time battery saver warning was the one causing
a change in settings global the radio button picker doesn't update
the preferences correctly and would put the seekbar at the top. This
explicitly adds an order to the seekbar so that if it gets updated
for some reason it will show up at the end again.

Test: atest BatterySaverScheduleSettingsTest
Bug: 131434758
Change-Id: Iebab5d176c8fbdaea078e13a71eff7e21408353b

Former-commit-id: bd6045bea92453d80ae2d878cd71ab63cfbb7a3e