Fix dark theme and battery saver interaction
In settings we weren't properly taking into account battery saver
enabled dark theme. This CL makes it so that when battery saver
on we disable the toggle until battery saver is off since
you can't change dark theme in battery saver. Additionally it
adds a message indicating why the toggle was disabled so the
user doesn't get confused.

Test: robotests
Bug: 131913864
Change-Id: I855e429a666098c837650852d1e1355477afa13d

Former-commit-id: ee5f79edbd3a1c05780de3e1bbfe0ccba0f605f4