Revert "Revert "Make subsettings display edge-to-edge""
This reverts commit 6d4a91a0dd665ed70a1409d1a4f9819581503982 [formerly 9e98b8cf0fd0669fdc3c351aafb91fa8e4a0d9ad].

We have to set clipToPadding false on scrollView to make it extend below
navbar. But the side effect is it stops clipping to padding. So it
overlaps with sibling views. In this case anything pinned on top of
screen (switchbar, pinned header for most UIs, and preview indicator for
font size page).

To fix them:
1. in preview pages, just revert back to clipTo*. Their UI doesn't scroll
   so it they don't need to special clip flag to begin with.
2. For pinned to top headers, we need to change the "z-order" by setting
   elevation and make sure background is not transparent.

Fixes: 132182711
Fixes: 132672399
Test: visually on normal screen, font size scree, wifi main setting UI,
      wifi data usage screen

Reason for revert: Rollforward with fix.

Change-Id: Ib6f7c4167b9a680a336f2a7310d82340094d3aac

Former-commit-id: b4b7725937337ae3677d98e0cb34adabc07856ce