Change switch text color to black in dark mode
It seems that the attribute values that are used to text color
in the switch bar are selectors which leads to problems because
they include an alpha value when the switch is "off". This would
cause the contrast between the text and the background to be too
low for GAR. This CL simply makes it so that in dark mode the text
for the switch bar is just true black since adjusting the
attributes would have wide scale implications for all UIs using
device default themes.

Test: visual inspection
Bug: 132755745
Change-Id: I309db82e9bf020fb654e0b93f6ec84380ae092f7

Former-commit-id: a2e7357535dabe66c3f1737d1755aa15a18d6ad7