Fix Settings search result 'add an account' does nothing.
Tap 'add an account' from search results will launch ChooseAccountFragment,
this way doesn't invoke AddAccountSettings and the add account service of gms
won't be triggered, so the items on the fragment will do nothing.

Move the key word to the search synonym of AccountDashboardFragment
and remove it from ChooseAccountFragment, the search result will invoke
AccountDashboardFragment.

Fixes: 133511289
Test: maunal
Change-Id: I67fa004c4965da37c59cada6ca27752e9fe7202b

Former-commit-id: 1903470bfb469df315aa3a83ba6dbed5cc40549f